Solution to Check Triangle Validity

+++ Check Triangle Validity@@6788dd63490637cfc44e1b93/code.c
@@ -4,8 +4,8 @@
 int main(){
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    int d=sqrt(a*a+b*b);
-    if (c==d){
+    float d=sqrt(a*a+b*b);
+    if (c==(int)d){
         printf("Valid");
     }
     else{
